+++ source/extensions/omni.isaac.gym.benchmark/omni/isaac/gym/benchmark/tests/test_benchmark_gym_standalone.py
@@ -14,7 +14,7 @@ import carb
 import omni.kit
 import omni.kit.test
 import omni.timeline
-from omni.isaac.benchmark.services.base_isaac_benchmark_async import BaseIsaacBenchmarkAsync
+from omni.isaac.benchmark.services import BaseIsaacBenchmarkAsync
 
 MAX_ITERATIONS = 10